refactor: reorganized file locations
