+++ lib/core/config/router.dart
@@ -6,9 +6,9 @@ import 'package:flex_workout_mobile/core/extensions/ui_extensions.dart';
 import 'package:flex_workout_mobile/features/auth/providers.dart';
 import 'package:flex_workout_mobile/features/auth/ui/screens/onboarding_screen.dart';
 import 'package:flex_workout_mobile/features/auth/ui/screens/profile_screen.dart';
-import 'package:flex_workout_mobile/features/exercise/ui/components/equipment_picker.dart';
-import 'package:flex_workout_mobile/features/exercise/ui/components/movement_pattern_picker.dart';
-import 'package:flex_workout_mobile/features/exercise/ui/components/primary_muscle_group_picker.dart';
+import 'package:flex_workout_mobile/features/exercise/ui/forms/equipment_picker.dart';
+import 'package:flex_workout_mobile/features/exercise/ui/forms/movement_pattern_picker.dart';
+import 'package:flex_workout_mobile/features/exercise/ui/forms/primary_muscle_group_picker.dart';
 import 'package:flex_workout_mobile/features/exercise/ui/screens/exercise_create_screen.dart';
 import 'package:flex_workout_mobile/features/exercise/ui/screens/exercise_view_screen.dart';
 import 'package:flex_workout_mobile/features/tracker/data/models/live_workout_model.dart';
