+++ lib/features/exercise/ui/containers/exercise_create_step_two.dart
@@ -7,9 +7,9 @@ import 'package:flex_workout_mobile/core/theme/app_layout.dart';
 import 'package:flex_workout_mobile/features/exercise/controllers/exercise_form_controller.dart';
 import 'package:flex_workout_mobile/features/exercise/data/models/exercise_form_model.dart';
 import 'package:flex_workout_mobile/features/exercise/data/models/exercise_model.dart';
-import 'package:flex_workout_mobile/features/exercise/ui/components/equipment_picker.dart';
-import 'package:flex_workout_mobile/features/exercise/ui/components/flex_picker.dart';
-import 'package:flex_workout_mobile/features/exercise/ui/components/movement_pattern_picker.dart';
+import 'package:flex_workout_mobile/features/exercise/ui/forms/equipment_picker.dart';
+import 'package:flex_workout_mobile/core/common/ui/forms/flex_picker.dart';
+import 'package:flex_workout_mobile/features/exercise/ui/forms/movement_pattern_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:material_symbols_icons/symbols.dart';
