+++ lib/features/exercise/ui/forms/equipment_picker.dart
